Added list styling + toolbar CSS class
Since the exact same flexbox styling will be used multiple times, I
moved it to a class. Have not yet decided whether this was a good idea.
Might remove it again if I end up tweaking the flexbox' configurations
for every toolbar anyway.

Styled the navigation toolbar (for sorting and searching) and the list
of items. Still to be done there: change the javascript accordingly, but
since this means changing the todo-item generation code that will be
changed anyway, I will do this later.