Buttons can be styled with an icon
Added the iconButton class to provide a quick and easy way to style a
button with an icon (it can now be set with one line of CSS).

A font with lots of icons was added as well. Only the font files are
used, the CSS files were added for completeness.