Added support for small screens
On small screens, everything will be displayed below eachother instead
of in two columns. 1200px was chosen as breakpoint because mobile phones
have a high screen resolution these days and (e.g. mine is 1920x1080,
while my laptop is 1600x900). The breakpoint should be between the
horizontal resolution of laptops and the horizontal resolution of phones
in portrait configuration. This is also taking into account the big UI
elements we have to accomodate elderly people.